[WIP] remove thin-pool lv if it is not used anymore

+++ executors/cmdexec/brick.go
@@ -11,6 +11,7 @@ package cmdexec
 
 import (
 	"fmt"
+	"strconv"
 	"strings"
 
 	"github.com/heketi/heketi/executors"
@@ -119,8 +120,8 @@ func (s *CmdExecutor) BrickDestroy(host string,
 		fmt.Sprintf("mount | grep -w %v | cut -d\" \" -f1", brick.Path),
 	}
 	output, err := s.RemoteExecutor.RemoteCommandExecute(host, commands, 5)
-	if err != nil {
-		logger.Err(err)
+	if output == nil || err != nil {
+		return fmt.Errorf("No brick mouned on %v, unable to proceed with removing", brick.Path)
 	}
 	dev := output[0]
 	// detect the thinp LV used by this brick (in "vg_.../tp_..." format)
@@ -151,28 +152,35 @@ func (s *CmdExecutor) BrickDestroy(host string,
 		logger.Err(err)
 	}
 
-	// Remove the thin-pool
-	// TODO: skip if there are more than one users of the tp
+	// Detect the number of bricks using the thin-pool
 	commands = []string{
-		fmt.Sprintf("lvremove -f %v", tp),
+		fmt.Sprintf("lvs --noheadings --options=thin_count %v", tp),
 	}
-	_, err = s.RemoteExecutor.RemoteCommandExecute(host, commands, 5)
+	output, err = s.RemoteExecutor.RemoteCommandExecute(host, commands, 5)
 	if err != nil {
 		logger.Err(err)
+		return fmt.Errorf("Unable to determine number of logical volumes in "+
+			"thin pool %v on host %v", tp, host)
+	}
+	thin_count, err := strconv.Atoi(strings.TrimSpace(output[0]))
+	if err != nil {
+		return fmt.Errorf("Failed to convert number of logical volumes in thin pool %v on host %v: %v", tp, host, err)
 	}
 
-	// Now cleanup the mount point
-	commands = []string{
-		fmt.Sprintf("rmdir %v", brick.Path),
-	}
-	_, err = s.RemoteExecutor.RemoteCommandExecute(host, commands, 5)
-	if err != nil {
-		logger.Err(err)
+	// If there is no brick left in the thin-pool, it can be removed
+	if thin_count == 0 {
+		commands = []string{
+			fmt.Sprintf("lvremove -f %v", tp),
+		}
+		_, err = s.RemoteExecutor.RemoteCommandExecute(host, commands, 5)
+		if err != nil {
+			logger.Err(err)
+		}
 	}
 
 	// Remove from fstab
 	// If the brick.Path contains "(/var)?/run/gluster/", there is no entry in fstab as GlusterD manages it.
-	if (!(strings.HasPrefix(brick.Path, "/run/gluster/") || strings.HasPrefix(brick.Path, "/var/run/gluster/"))) {
+	if !(strings.HasPrefix(brick.Path, "/run/gluster/") || strings.HasPrefix(brick.Path, "/var/run/gluster/")) {
 		commands = []string{
 			fmt.Sprintf("sed -i.save \"/%v/d\" %v",
 				utils.BrickIdToName(brick.Name),
